Update GenerateNewTerrainList script
No longer passes directly reference to GlobalInformation to the TerrainObjects Information variable. This means I can no longer live update the terrains from the global manager via the inspector, but I does allow the groundwork to be laid for passing the global information that's been passed as a new declared variable within the script in which the values will be inteneded to be modified to help keep terrain smooth continue from their intended spaced out positions.