Fix issue with mesh seams - largly fixed
Starting debugging the issue in regards to the terrain generator, and managed to find a fix, albeit poorly understood, that largely solved the issue in regards to the different terrains not meshing well together – I had an issue in regards to the Pelin Offsets being calculated incorrectly

The fix below shows that the number has some significance in the calculation as my terrains now mesh together with the perlin offset and perlin scale values being changed. Needs further investigation.