Cleaning up the code and comments.
Spent time commenting up the code and applying tool-tips to the TerrainInformation class and lastly removing unused and / or legacy features. Need to re-look at where the randomisation of System.Random(seed) is being applied – is this producing issues with seams or not?). Ideally need to re-work the seed application to the global information script.