Add animation curve and fix duplication of terrain via manager, fix perlin offsetting of adjacent terrains
I resolved the bug that was causing duplication of terrain when modifying the terrain information in the manager class, which was subsequently applied to the TerraianObjects themselves. The fix was simple – the list of terrain objects was changed to a static variable.

Furthermore, I fixed the longstanding bug with terrain perlin offsets – now the meshes maintain a reasonable level of similarity at the edges. The key was in dividing the positional offset value of a given axis by the grid spacing in order to get the positional offset normalised. This value was then divided by the PerlinScale of the center terrain position to retrieve the value that would need to be added to the PerlinOffset for the axis of the offset terrain.

I added an animation curve evaluation to the Y values generated by Perlin Noise that are applied to the heigh of the mesh – this allows given ranges to be flattened or curves / sloped for the sake of allowing greater customization of the terrain types being produced.